+++ code/cell.electron/app/src/main/main.sys/sys.define.types.ts
