+++ code/cell.electron/app/src/main/main.sys/sys.ipc.ts
@@ -1,15 +1,13 @@
 import { debounceTime } from 'rxjs/operators';
-
 import { constants, t } from '../common';
-
-const { IPC } = constants;
 
 /**
  * Initializes the IPC (inter-process communications) event streams.
  */
-export function ipc(args: { ctx: t.IAppCtx__OLD }) {
+export function ipc(args: { ctx: t.IContext }) {
   const { ctx } = args;
   const { client } = ctx;
+  const { IPC } = constants;
 
   // Broadcast changes to each window.
   client.changes.changed$.pipe(debounceTime(50)).subscribe(e => {
