+++ code/cell.electron/app/src/main/main.sys/sys.ts
@@ -1,6 +1,1 @@
-import '../../config';
-
 export * from './sys.init';
-export * from './sys.upload';
-export * from './sys.typeDefs';
-export * from './sys.typeDefs.init';
