--- code/cell.electron/app/src/main/main.sys/sys.typeDefs.ts
@@ -1,33 +1,0 @@
-import { constants, TypeSystem } from '../common';
-
-const NS = constants.SYS.NS;
-const def = TypeSystem.def();
-
-def
-  .ns(NS.TYPE.APP)
-  .type('SysApp')
-  .prop('title', { type: 'string', default: 'CellOS' })
-  .prop('windowDefs', p => p.type('/SysAppWindowDef[]').target('ref'))
-  .prop('windows', p => p.type('/SysAppWindow[]').target('ref'));
-
-def
-  .ns(NS.TYPE.WINDOW_DEF)
-  .type('SysAppWindowDef')
-  .prop('kind', p => p.type('string').default(''))
-  .prop('width', p => p.type('number').default(1200))
-  .prop('height', p => p.type('number').default(800))
-  .prop('sandbox', p => p.type('boolean').default(true))
-  .prop('backgroundColor', p => p.type('string').default('#fff'));
-
-def
-  .ns(NS.TYPE.WINDOW)
-  .type('SysAppWindow')
-  .prop('id', p => p.type('string').default(''))
-  .prop('kind', p => p.type('string').default(''))
-  .prop('title', p => p.type('string').default('Untitled'))
-  .prop('width', p => p.type('number').default(-1))
-  .prop('height', p => p.type('number').default(-1))
-  .prop('x', 'number')
-  .prop('y', 'number');
-
-export const DEFS = def.toObject();
