--- code/cell.electron/app/src/main/main.sys2/index.ts
@@ -1,2 +1,0 @@
-import * as sys from './sys';
-export { sys };
