--- code/cell.electron/app/src/main/main.sys2/sys.ts
@@ -1,1 +1,0 @@
-export * from './sys.init';
