+++ code/cell.electron/app/src/main/main.tray/tray.ts
@@ -1,8 +1,8 @@
 import { Tray, Menu } from 'electron';
 import { fs, t } from '../common';
-import { createWindow } from '../main.window';
+// import { createWindow } from '../main.window';
 
-export function init(args: { host: string; def: string; ctx: t.IAppCtx__OLD }) {
+export function init(args: { host: string; def: string; ctx: t.IContext }) {
   const { host, def, ctx } = args;
 
   const icon = fs.join(__dirname, '../../../assets/icons/tray/tray.png');
