+++ code/cell.electron/app/src/main/main.window/index.ts
@@ -1,1 +1,2 @@
-export * from './window';
+import * as window from './window';
+export { window };
