+++ code/cell.electron/app/src/main/main.window/window.log.ts
