--- code/cell.electron/app/src/main/main.window2/index.ts
@@ -1,2 +1,0 @@
-import * as window from './window';
-export { window };
