--- code/cell.electron/app/src/main/main.window2/window.ts
@@ -1,1 +1,0 @@
-export * from './window.create';
